Allow connections crossing the north axis
Previously, Redrawn-Space-Connections saw 0.9 turns and 0.1 turns as 0.8 turns away. This corrects the formula to instead give 0.2 turns.

This is relevant for Igrys, which (after offsetting from Vulcanus) had an angle of around 0.93 turns, making Redrawn-Space-Connections think it was very far away from the other planets, and made a single connection all the way to Maraxsis on the other side of the star.